pyproject.toml
[project]
name = "llm-summarize"
version = "0.1.0"
description = "LLM同士のチャット実験アプリケーション"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.32.0",
    "google-genai>=1.3.0",
    "openai>=1.12.0",
    "anthropic>=0.8.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.1",
]
